Impl bindings for receiver, sender
We don't want to expose the low-level `frame` structure, so both receiver and
sender use functions like `ReadBytes` and `WriteBytes` that operate on []byte.

I also added a simple "integration test" that writes something with sender,
reads it with receiver and checks that what we received is not an array of 0s.
Similar smoke-test can be found
[here](https://github.com/roc-project/roc-tests/blob/master/tests/test_service_quality.cpp).